:recycle: [refactor] : menu 함수 custom hook으로 분리
